Ignoring produced change files (see tools/release.sh).
In order to have a minimum of data and, possibly, a partition of
changes to web sources, the change files are assembled from elementary
shared chunks. This is done by the NTS team webmerge utility (and
needs a working tex engine). This is done invoking tools/release.sh
and the produced files have to be ignored.