Reverting CWEB to the last version of D.E. Knuth and Silvio Levy.
The new version, maintained not from the original authors, present
regressions: included files (that I use a lot) have no comment
generated in C files leading to failure to use a debugger correctly;
secondly, for formatting, cweave depends on a file not provided,
present in TeXlive based distributions and part of the LaTeX project.

No. Back to normal.