README.md: clarify that the module needs external sources.
KerTeX_T has everything needed (except for the translation of WEB into
C: that is kerTeX_M) to build the distribution. But here are only
sources maintained by T. Laronde. For example the D.E. Knuth sources
are not here and have to be retrieved (a bundle is at disposal for
download on http://kertex.kergis.com/ with all the external sources
used). So state it clearly in the README.