mp2ps: correct order of pages in produced document.
The \shipout needs to be expressed for every vbox. Omitting it for
the first figure was defining a vbox, not printed (while the others
were because shipout expects a vbox) and thus was rendered last and
not first.

XXX There is a blank page at end, because of \end, and I will have to
look if it is possible to flush and terminate correctly the file
without creating this unwanted artefact.